Updated to v1.0.0-BETA3
Fixed /is help command from having different colors, which never looked right, between the / - is command, so it would come up as: AQUA: / GREEN: is. We've now made it so it has: YELLOW: / YELLOW: /is.
Added more new aliases to commands. Such as:
/is make = /is create
/is clear = /is reset
/is ourmembers, and /is listmembers = /is members.
/is tpto = /is tp
/is ask = /is invite

Added back /is join:
/is join = go
/is join is now a aliases command, with /is go.
Updated to V1.0.0-BETA3