Updated to v1.0.4
Fixed /is
Added new commands: /is aliases, /is todolist, /is info, /is faq, and /is howtoplay.
But be warned, that these commands currently do not function. These are the commands that have been added to the /is help command / page for show. But we'll add these commands very soon. Enjoy.